fix(ci): line-limit maintainability — split oversized remote repair helpers tests
